streamlit_app_2/requirements.txt
streamlit==1.46.1
google-genai==1.24.0
plaid-python==35.0.0
